JHy updated DrainageSystem-1-2
After reviewing the file against the original LandXML, noticed that the sizes of pipes and manholes don’t match. The diameters given in LandXML for Structs and Pipes seem to have been taken as outer dimensions, when they are inner values instead. This doesn’t really matter in unit tests, but for a good example to pass around, I’d like to get it fixed.

Another small issue is the Axis representation, which now follows the pipe centre-lines, but the domain practice (in Finland, anyway) is to follow the invert level for gravity-fed lines (outside-top level for pressure-pipes).

I’ve made the changes in the profile definitions to get the sizes right, and dealt with the Axis-issue by changing the local placements of the pipes to have origin at invert level (each axis polyline starting at 0,0,0), and indicating the used cardinal point by changing the value in IFCMATERIALPROFILESETUSAGE from 5 to 20 (see below), and also giving appropriate offset in IFCCIRCLEHOLLOWPROFILEDEF.Position.